pyproject.toml
[tool.poetry]
name = "finance-analysis"
version = "1.0.0"
description = ""
authors = ["Author Name <tomerroditi1@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "==2.2.3"
numpy = "==1.26.4"
pyarrow = "==15.0.2"
plotly = "==5.24.1"
PyYAML = "==6.0.1"
SQLAlchemy = "==2.0.29"
keyring = ">=24.0.0"
# Backend API dependencies
fastapi = ">=0.109.0"
uvicorn = { version = ">=0.27.0", extras = ["standard"] }
python-dotenv = "^1.0.0"
alembic = "^1.14.0"
# Scraper dependencies
playwright = "^1.57.0"
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "==8.3.5"
Faker = "==24.3.0"
commitizen = "^4.4.1"
playwright = "^1.57.0"
requests = "^2.32.5"
ruff = "^0.9.0"
pytest-cov = "^7.0.0"

[tool.pytest.ini_options]
pythonpath = ["."]
markers = ["sensitive: marks tests that pulls sensitive data from the internet"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver2"
version_provider = "poetry"
version_files = [
    "pyproject.toml:tool.poetry.version",
    "frontend/package.json:\"version\"",
    "build/installer_script.nsi:^!define APP_VERSION \".*\"$",
]
update_changelog_on_bump = true
bump_message = "bump: version $current_version → $new_version [skip ci]"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
